pyproject.toml
[tool.poetry]
name = "hackrads"
version = "0.1.0"
description = ""
authors = ["Marcus Paredes <viniciusparedes@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.72.0"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
